requirements.txt
django==1.10.2
djangorestframework==3.5.3
django-filter==0.15.3
django-rest-swagger==2.0.7
django-crispy-forms==1.6.1
django-configurations==2.0
django-gravatar2==1.4.0
django-registration==2.1.2
djangorestframework-filters==0.9.0
django-oauth-toolkit==0.10.0
django-rest-assured==0.2.0
django-polymorphic==1.0.2
django-cors-headers==1.3.0
factory-boy==2.7.0
Faker==0.7.3
humanize==0.5.1
aiohttp-wsgi==0.6.3
coreapi==2.0.8
#pyjade==4.0.0
https://github.com/syrusakbary/pyjade/archive/master.zip
django-settings-export==1.2.1
django-auth-ldap==1.2.8
